+++ library/Hakyll/Images/Internal.hs
@@ -1,0 +1,16 @@
+-- |
+-- Module      : Hakyll.Images.Internal
+-- Description : Hakyll utilities for image files
+-- Copyright   : (c) Laurent P René de Cotret, 2019 - present
+-- License     : BSD3
+-- Maintainer  : laurent.decotret@outlook.com
+-- Stability   : unstable
+-- Portability : portable
+--
+-- Internal re-exports used for tests.
+module Hakyll.Images.Internal
+  ( module Hakyll.Images.Common,
+  )
+where
+
+import Hakyll.Images.Common