+++ library/Hakyll/Images/Resize.hs
@@ -46,31 +46,25 @@ module Hakyll.Images.Resize
   )
 where
 
-import Codec.Picture (convertRGBA8, decodeImage)
+import Codec.Picture (convertRGBA8)
 import Codec.Picture.Extra (scaleBilinear)
 import Codec.Picture.Types (DynamicImage (..), imageHeight, imageWidth)
-import Data.ByteString (ByteString)
 import Data.Ratio ((%))
 import Hakyll.Core.Compiler (Compiler)
 import Hakyll.Core.Item (Item (..))
-import Hakyll.Images.Common (Image (..), encode)
+import Hakyll.Images.Common (Image (..), ImageContent, WithMetadata (..), encode, withImageContent)
 
 type Width = Int
 
 type Height = Int
-
-decodeImage' :: ByteString -> DynamicImage
-decodeImage' im = case decodeImage im of
-  Left msg -> error msg
-  Right im' -> im'
 
 -- | Resize an image to specified width and height using the bilinear transform.
 -- The aspect ratio may not be respected.
 --
 -- In the process, an image is converted to RGBA8. Therefore, some information
 -- loss may occur.
-resize :: Width -> Height -> DynamicImage -> DynamicImage
-resize w h = ImageRGBA8 . scaleBilinear w h . convertRGBA8
+resize :: Width -> Height -> ImageContent -> ImageContent
+resize w h = fmap (ImageRGBA8 . scaleBilinear w h . convertRGBA8)
 
 -- | Scale an image to a size that will fit in the specified width and height,
 -- while preserving aspect ratio. Images might be scaled up as well.
@@ -79,7 +73,7 @@ resize w h = ImageRGBA8 . scaleBilinear w h . convertRGBA8
 -- loss may occur.
 --
 -- To scale images down only, take a look at 'ensureFit'.
-scale :: Width -> Height -> DynamicImage -> DynamicImage
+scale :: Width -> Height -> ImageContent -> ImageContent
 scale w h = scale' w h True
 
 -- | Scale an image to a size that will fit in the specified width and height,
@@ -95,10 +89,10 @@ scale' ::
   -- | Allow scaling up as well.
   Bool ->
   -- | Source image
-  DynamicImage ->
+  ImageContent ->
   -- | Scaled image.
-  DynamicImage
-scale' w h upAllowed img = resize maxWidth maxHeight img
+  ImageContent
+scale' w h upAllowed content@(MkWithMetadata img _) = resize maxWidth maxHeight content
   where
     img' = convertRGBA8 img -- Required to extract height and width
     (imgWidth, imgHeight) = (imageWidth img', imageHeight img')
@@ -119,7 +113,7 @@ scale' w h upAllowed img = resize maxWidth maxHeight img
 -- loss may occur.
 --
 -- To scale images up __or__ down, take a look at 'scale'.
-ensureFit :: Width -> Height -> DynamicImage -> DynamicImage
+ensureFit :: Width -> Height -> ImageContent -> ImageContent
 ensureFit w h = scale' w h False
 
 -- | Compiler that resizes images to a specific dimensions. Aspect ratio
@@ -127,45 +121,39 @@ ensureFit w h = scale' w h False
 --
 -- @
 -- match "*.png" $ do
---     route idRoute
---     compile $ loadImage >>= resizeImageCompiler 48 64
+--    route idRoute
+--    compile $ loadImage >>= resizeImageCompiler 48 64
 -- @
 --
 -- Note that in the resizing process, images will be converted to RGBA8.
 -- To preserve aspect ratio, take a look at 'scaleImageCompiler'.
 resizeImageCompiler :: Width -> Height -> Item Image -> Compiler (Item Image)
-resizeImageCompiler w h item =
-  let fmt = (format . itemBody) item
-   in return $ encode fmt . resize w h . decodeImage' . image <$> item
+resizeImageCompiler w h = return . fmap (withImageContent (resize w h) encode)
 
 -- | Compiler that rescales images to fit within dimensions. Aspect ratio
 -- will be preserved. Images might be scaled up as well.
 --
 -- @
 -- match "*.tiff" $ do
---     route idRoute
---     compile $ loadImage >>= scaleImageCompiler 48 64
+--    route idRoute
+--    compile $ loadImage >>= scaleImageCompiler 48 64
 -- @
 --
 -- Note that in the resizing process, images will be converted to RGBA8.
 -- To ensure images are only scaled __down__, take a look at 'ensureFitCompiler'.
 scaleImageCompiler :: Width -> Height -> Item Image -> Compiler (Item Image)
-scaleImageCompiler w h item =
-  let fmt = (format . itemBody) item
-   in return $ encode fmt . scale w h . decodeImage' . image <$> item
+scaleImageCompiler w h = return . fmap (withImageContent (scale w h) encode)
 
 -- | Compiler that ensures images will fit within dimensions. Images might
 -- be scaled down, but never up.  Aspect ratio will be preserved.
 --
 -- @
 -- match "*.tiff" $ do
---     route idRoute
---     compile $ loadImage >>= ensureFitCompiler 48 64
+--    route idRoute
+--    compile $ loadImage >>= ensureFitCompiler 48 64
 -- @
 --
 -- Note that in the resizing process, images will be converted to RGBA8.
 -- To allow the possibility of scaling up, take a look at 'scaleImageCompiler'.
 ensureFitCompiler :: Width -> Height -> Item Image -> Compiler (Item Image)
-ensureFitCompiler w h item =
-  let fmt = (format . itemBody) item
-   in return $ encode fmt . ensureFit w h . decodeImage' . image <$> item
+ensureFitCompiler w h = return . fmap (withImageContent (ensureFit w h) encode)