+++ tests/TestSuite.hs
@@ -12,11 +12,12 @@ import qualified Hakyll.Images.Common.Tests
 import qualified Hakyll.Images.CompressJpg.Tests
 import qualified Hakyll.Images.Metadata.Tests
 import qualified Hakyll.Images.Resize.Tests
+import qualified Hakyll.Images.Tests.Utils
 import Test.Tasty (defaultMain, testGroup)
 
 --------------------------------------------------------------------------------
 main :: IO ()
-main =
+main = do
   defaultMain $
     testGroup
       "Hakyll"
@@ -25,3 +26,5 @@ main =
         Hakyll.Images.Resize.Tests.tests,
         Hakyll.Images.Metadata.Tests.tests
       ]
+
+  Hakyll.Images.Tests.Utils.cleanTestEnv